various small cleanups
